DataAbstractionLayer/postgres_db.py
import logging
from typing import Optional, List

# SQLAlchemy es el ORM estándar en el ecosistema Python.
# asyncpg es el driver asíncrono para PostgreSQL. Ambas dependencias ya están en requirements.txt.
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
from sqlalchemy.orm import sessionmaker
from contextlib import asynccontextmanager

from dal.base import BaseDatabase
from schemas.db_models import MachineModel, OrderModel, ProductionQueueItem
from core.config import settings

logger = logging.getLogger(__name__)

class PostgresDB(BaseDatabase):
    """
    Implementación de la capa de abstracción de datos para una base de datos
    PostgreSQL compartida, utilizando SQLAlchemy para la comunicación asíncrona.
    """
    def __init__(self, database_url: str):
        # El motor de la base de datos se crea una vez.
        self.engine = create_async_engine(database_url, echo=False)
        # La fábrica de sesiones se configura una vez.
        self.async_session = sessionmaker(
            self.engine, expire_on_commit=False, class_=AsyncSession
        )
        logger.info("Implementación de DAL para PostgreSQL inicializada.")

    async def get_machine_by_id(self, machine_id: str) -> Optional[MachineModel]:
        """Obtiene una máquina de la base de datos real."""
        async with self.async_session() as session:
            # Aquí iría tu lógica de SQLAlchemy para consultar la tabla 'machines'.
            # Ejemplo (requiere definir los modelos de SQLAlchemy):
            # result = await session.execute(select(Machine).where(Machine.id == machine_id))
            # machine_db = result.scalar_one_or_none()
            # return MachineModel.from_orm(machine_db) if machine_db else None
            pass # Placeholder
    
    async def update_machine_status(self, machine_id: str, status: str, disabled_units: Optional[int] = None) -> bool:
        """Actualiza el estado de una máquina en la base de datos real."""
        async with self.async_session() as session:
            # Aquí iría tu lógica para actualizar un registro y hacer commit.
            # Ejemplo:
            # await session.execute(update(Machine).where(Machine.id == machine_id).values(status=status))
            # await session.commit()
            return True # Placeholder
    
    # ... Aquí implementarías el resto de los métodos abstractos de BaseDatabase
    # utilizando sesiones de SQLAlchemy para interactuar con la DB compartida.
    async def get_order_by_id(self, order_id: str) -> Optional[OrderModel]:
        pass
        
    async def get_production_queue(self, machine_id: Optional[str] = None) -> List[ProductionQueueItem]:
        return []

    async def prioritize_order(self, order_id: str, machine_id: Optional[str] = None, reorder_rest: bool = True) -> bool:
        return True

    async def register_roll_weight(self, machine_id: str, order_id: str, weight_kg: float) -> bool:
        return True

    async def register_waste(self, order_id: str, machine_id: str, weight_kg: float, reason: str) -> bool:
        return True

    async def get_all_machines_status(self) -> List[MachineModel]:
        return []

    @asynccontextmanager
    async def get_session(self):
        conn = await self.pool.acquire()
        try:
            yield conn
        finally:
            await conn.release()